Applied permission check in updateUser and deleteUser

+++ routes/users.js
@@ -63,13 +63,26 @@ async function createUser(ctx) {
     }
 }
 
-async function updateUser(ctx){
+async function updateUser(ctx) {
     let id = ctx.params.id;
-    let body = ctx.request.body;
-    let result = await model.updateById(body, id);
-    if (result.affectedRows) {
-        ctx.body = {ID: id, updated: true};
-        ctx.status = 200;
+    let result = await model.getById(id);
+  
+    if (result.length) {
+        let data = result[0];
+        let permission = can.update(ctx.state.user, data);
+
+        if (!permission.granted) {
+            ctx.status = 403;
+        } else {
+            let newData = permission.filter(ctx.request.body);
+            result = await model.updateById(newData, id);
+            if (result.affectedRows) {
+                ctx.body = {ID: id, updated: true};
+                ctx.status = 200;
+            } else {
+                ctx.status = 404;
+            }
+        }
     } else {
         ctx.status = 404;
     }
@@ -77,10 +90,23 @@ async function updateUser(ctx){
 
 async function deleteUser(ctx){
     let id = ctx.params.id;
-    let result = await model.deleteById(id);
-    if (result.affectedRows) {
-        ctx.body = {ID: id, deleted: true};
-        ctx.status = 200;
+    let result = await model.getById(id);
+  
+    if (result.length) {
+        let data = result[0];
+        let permission = can.delete(ctx.state.user, data);
+  
+        if (!permission.granted) {
+            ctx.status = 403;
+        } else {
+            let result = await model.deleteById(id);
+            if (result.affectedRows) {
+                ctx.body = {ID: id, deleted: true};
+                ctx.status = 200;
+            } else {
+                ctx.status = 404;
+            }
+        }    
     } else {
         ctx.status = 404;
     }
